Adds artist-mode support to spacemacs
`artist-mode` allows you to draw ASCII things on your text, but it
requires you to enter emacs mode.

This commit adds a hook to enter automatically emacs mode on artist
enablement, and a a shortcut on the spacemacs leader to toggle it.

`SPC a a` or `M-m a a` to enter and exit it.